requirements.txt
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
filelock==3.18.0
fonttools @ file:///croot/fonttools_1737039080035/work
fsspec==2025.3.2
Jinja2==3.1.6
joblib @ file:///work/perseverance-python-buildout/croot/joblib_1728392999985/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
MarkupSafe==3.0.2
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
networkx==3.4.2
numexpr @ file:///io/perseverance-python-buildout/croot/numexpr_1731701292781/work
numpy @ file:///croot/numpy_and_numpy_base_1747238018250/work/dist/numpy-2.2.5-cp313-cp313-linux_x86_64.whl#sha256=0e0751551faf5084bae97319f1019238122ff4a3eefee4d271cdc73fe7ee20eb
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp313-cp313-linux_x86_64.whl#sha256=248baa16b8f7b48e3a9c0e641fba9a72bce9c95c0dddf7c9a7547016b61fe3fb
pillow @ file:///croot/pillow_1744613067434/work
pyparsing @ file:///io/buildout/croot/pyparsing_1735862672222/work
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1744804475988/work/pyqt_sip
python-dateutil @ file:///work/perseverance-python-buildout/croot/python-dateutil_1728385355105/work
pytz @ file:///work/perseverance-python-buildout/croot/pytz_1728386814080/work
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1747238027288/work/dist/scipy-1.15.3-cp313-cp313-linux_x86_64.whl#sha256=ef0f7946c766acf9935b91f06cb7735ddb4890c85109a813e5ea260749c68c5c
setuptools==72.1.0
sip @ file:///croot/sip_1738856193618/work
six @ file:///croot/six_1744271502820/work
sympy==1.14.0
threadpoolctl @ file:///croot/threadpoolctl_1731006087558/work
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado @ file:///io/buildout/croot/tornado_1735842129908/work
triton==3.3.0
typing_extensions==4.13.2
tzdata @ file:///croot/python-tzdata_1746123641790/work
